Moves most role checks to DB operation

+++ api/src/controllers/document-controller.js
@@ -7,8 +7,6 @@ const queryUtils = require('../utils/query-utils');
 const businessLogicManager = require('../utils/business-logic-manager');
 const mongodb = require('../utils/mongodb');
 const defaultLog = require('../utils/logger')('record');
-const { userInRole } = require('../utils/auth-utils');
-const { ROLES } = require('../utils/constants/misc');
 
 const OBJ_STORE_URL = process.env.OBJECT_STORE_endpoint_url || 'nrs.objectstore.gov.bc.ca';
 const ep = new AWS.Endpoint(OBJ_STORE_URL);
@@ -25,10 +23,6 @@ exports.protectedOptions = function(args, res, next) {
 };
 
 exports.protectedPost = async function(args, res, next) { // Confirm user has correct role.
-  if (!userInRole(ROLES.ADMIN_ROLES, args.swagger.params.auth_payload.realm_access.roles)) {
-    return queryActions.sendResponse(res, 400, 'Missing valid user role.');
-  }  
-
   if (
     args.swagger.params.fileName &&
     args.swagger.params.fileName.value &&
@@ -39,7 +33,7 @@ exports.protectedPost = async function(args, res, next) { // Confirm user has co
     const collection = db.collection('nrpti');
 
     // fetch master record
-    const masterRecord = await collection.findOne({ _id: new ObjectID(args.swagger.params.recordId.value) });
+    const masterRecord = await collection.findOne({ _id: new ObjectID(args.swagger.params.recordId.value), write: { $in: args.swagger.params.auth_payload.realm_access.roles } });
 
     // Set mongo document and s3 document roles
     const readRoles = [];
@@ -93,7 +87,7 @@ exports.protectedPost = async function(args, res, next) { // Confirm user has co
     try {
       // add to master record
       recordResponse = await collection.findOneAndUpdate(
-        { _id: new ObjectID(args.swagger.params.recordId.value) },
+        { _id: new ObjectID(args.swagger.params.recordId.value), write: { $in: args.swagger.params.auth_payload.realm_access.roles } },
         { $addToSet: { documents: new ObjectID(docResponse._id) } },
         { new: true }
       );
@@ -114,7 +108,7 @@ exports.protectedPost = async function(args, res, next) { // Confirm user has co
       recordResponse.value._flavourRecords.forEach(id => {
         observables.push(
           collection.findOneAndUpdate(
-            { _id: new ObjectID(id) },
+            { _id: new ObjectID(id), write: { $in: args.swagger.params.auth_payload.realm_access.roles } },
             { $addToSet: { documents: new ObjectID(docResponse._id) } },
             { new: true }
           )
@@ -143,10 +137,6 @@ exports.protectedPost = async function(args, res, next) { // Confirm user has co
 };
 
 exports.protectedDelete = async function(args, res, next) {
-  if (!userInRole(ROLES.ADMIN_ROLES, args.swagger.params.auth_payload.realm_access.roles)) {
-    return queryActions.sendResponse(res, 400, 'Missing valid user role.');
-  }
-
   if (
     args.swagger.params.docId &&
     args.swagger.params.docId.value &&
@@ -158,7 +148,7 @@ exports.protectedDelete = async function(args, res, next) {
     let docResponse = null;
     let s3Response = null;
     try {
-      docResponse = await Document.findOneAndRemove({ _id: new ObjectID(args.swagger.params.docId.value) });
+      docResponse = await Document.findOneAndRemove({ _id: new ObjectID(args.swagger.params.docId.value), write: { $in: args.swagger.params.auth_payload.realm_access.roles } });
     } catch (e) {
       defaultLog.info(`Error removing document meta ${args.swagger.params.docId.value} from the database: ${e}`);
       return queryActions.sendResponse(
@@ -193,7 +183,7 @@ exports.protectedDelete = async function(args, res, next) {
     try {
       // remove from master record
       recordResponse = await collection.findOneAndUpdate(
-        { _id: new ObjectID(args.swagger.params.recordId.value) },
+        { _id: new ObjectID(args.swagger.params.recordId.value), write: { $in: args.swagger.params.auth_payload.realm_access.roles } },
         { $pull: { documents: new ObjectID(docResponse._id) } },
         { new: true }
       );
@@ -214,7 +204,7 @@ exports.protectedDelete = async function(args, res, next) {
       recordResponse.value._flavourRecords.forEach(id => {
         observables.push(
           collection.findOneAndUpdate(
-            { _id: new ObjectID(id) },
+            { _id: new ObjectID(id), write: { $in: args.swagger.params.auth_payload.realm_access.roles } },
             { $pull: { documents: new ObjectID(docResponse._id) } },
             { new: true }
           )
@@ -400,7 +390,7 @@ async function publishDocument(docId, auth_payload) {
   }
 
   const Document = require('mongoose').model('Document');
-  const document = await Document.findOne({ _id: new ObjectID(docId) });
+  const document = await Document.findOne({ _id: new ObjectID(docId), write: { $in: auth_payload.realm_access.roles } });
 
   if (!document) {
     defaultLog.info(`publishDocument - couldn't find document for docId: ${docId}`);
@@ -429,7 +419,7 @@ async function unpublishDocument(docId, auth_payload) {
   }
 
   const Document = require('mongoose').model('Document');
-  const document = await Document.findOne({ _id: new ObjectID(docId) });
+  const document = await Document.findOne({ _id: new ObjectID(docId), write: { $in: auth_payload.realm_access.roles } });
 
   if (!document) {
     defaultLog.info(`unpublishDocument - couldn't find document for docId: ${docId}`);
@@ -453,17 +443,13 @@ exports.unpublishDocument = unpublishDocument;
  * @returns
  */
 exports.protectedGetS3SignedURL = async function(args, res, next) {
-  if (!userInRole(ROLES.ADMIN_ROLES, args.swagger.params.auth_payload.realm_access.roles)) {
-    return queryActions.sendResponse(res, 400, 'Missing valid user role.');
-  }  
-  
   if (!args.swagger.params.docId || !args.swagger.params.docId.value) {
     defaultLog.warn('protectedGet - missing required docId param');
     return queryActions.sendResponse(res, 400, 'Missing required docId param');
   }
 
   const Document = mongoose.model('Document');
-  const document = await Document.findOne({ _id: new ObjectID(args.swagger.params.docId.value) });
+  const document = await Document.findOne({ _id: new ObjectID(args.swagger.params.docId.value), write: { $in: args.swagger.params.auth_payload.realm_access.roles } });
 
   if (!document) {
     defaultLog.info(`protectedGetS3SignedURL - couldn't find document for docId: ${args.swagger.params.docId.value}`);
